Modified Login View
This error doesn't occur with simple view but only with partial view ( from my experience ). Login View was giving error 'resource not found' as it could not locate [HttpPost] action in my controller as it gave up finding it in Login.cshtml.
Adding "@Html.AntiForgeryToken()" in View solves the error.